fixed homepage bug
Homepage now successfully redirects to Search Results instead of crashing
