+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,10 @@
+{
+  "ExpandedNodes": [
+    "",
+    "\\HaloMMSiteMVC",
+    "\\HaloMMSiteMVC\\App_Start",
+    "\\HaloMMSiteMVC\\Controllers"
+  ],
+  "SelectedNode": "\\HaloMMSiteMVC.sln",
+  "PreviewInSolutionExplorer": false
+}
