+++ HaloMMSiteMVC/App_Start/BundleConfig.cs
@@ -25,7 +25,7 @@ namespace HaloMMSiteMVC
 
             bundles.Add(new StyleBundle("~/Content/css").Include(
                       "~/Content/bootstrap.css",
-                      "~/Content/site.css"));
+                      "~/Content/Site.css"));
         }
     }
 }
