+++ HaloMMSiteMVC/App_Start/RouteConfig.cs
@@ -16,7 +16,7 @@ namespace HaloMMSiteMVC
             routes.MapRoute(
                 name: "Default",
                 url: "{controller}/{action}/{id}",
-                defaults: new { controller = "Player", action = "Player", id = UrlParameter.Optional }
+                defaults: new { controller = "Home", action = "Player", id = UrlParameter.Optional }
             );
         }
     }
