+++ HaloMMSiteMVC/Content/Site.css
@@ -7,12 +7,13 @@
 .body-content {
     padding-left: 15px;
     padding-right: 15px;
+    
+    
 }
-
-/* Override the default bootstrap behavior where horizontal description lists 
+    /* Override the default bootstrap behavior where horizontal description lists 
    will truncate terms that are too long to fit in the left column 
 */
-.dl-horizontal dt {
+    .dl-horizontal dt {
     white-space: normal;
 }
 
@@ -22,3 +23,5 @@ select,
 textarea {
     max-width: 280px;
 }
+
+
