+++ HaloMMSiteMVC/Content/bootstrap.css
@@ -58,7 +58,8 @@ html {
 }
 
 body {
-  margin: 0;
+    margin: 0;
+    
 }
 
 a:focus {
@@ -224,6 +225,8 @@ table {
   border-spacing: 0;
 }
 
+
+
 @media print {
   * {
     color: #000 !important;
@@ -311,11 +314,12 @@ html {
 }
 
 body {
-  font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
-  font-size: 14px;
-  line-height: 1.428571429;
-  color: #333333;
-  background-color: #ffffff;
+    font-family: "Helvetica Neue", Helvetica, Arial, sans-serif;
+    font-size: 14px;
+    line-height: 1.428571429;
+    color: #333333;
+    background-color: #ffffff;
+  
 }
 
 input,
