--- HaloMMSiteMVC/Views/Home/About.cshtml
@@ -1,7 +1,0 @@
-﻿@{
-    ViewBag.Title = "About";
-}
-<h2>@ViewBag.Title.</h2>
-<h3>@ViewBag.Message</h3>
-
-<p>Use this area to provide additional information.</p>
